+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Changelog
 
-## 1.1.0 August xx, 2023
+## 1.2.0 August 18, 2023
+
+New API calls:
+- scrollToPixel(pixel:number[,behavior:string])
+- scrollByPixel(pixel:number[,behavior;string])
+
+Internal updates:
+- scroll mechanics for variable content was changed to accomodate scroll API calls. 
+
+## 1.1.0 August 13, 2023
 
 Breaking change:
 - `changeListSizeCallback` API replaces changeListsizeCallback (note the camel case)